refactor: 교회 소개 페이지 도메인 재정립
