+++ src/features/about/index.ts
@@ -1,0 +1,2 @@
+export { default as ChurchVision } from "./ui/ChurchVision";
+export { default as ChurchHistory } from "./ui/ChurchHistory";
