+++ src/features/churchInfo/index.ts
@@ -1,7 +1,5 @@
 export { default as WorshipSchedule } from "./ui/WorshipSchedule";
 export { default as ChurchIntro } from "./ui/ChurchIntro";
 export { default as HeroSection } from "./ui/HeroSection";
-export { default as ChurchVision } from "./ui/ChurchVision";
-export { default as PastorIntroduction } from "./ui/PastorIntroduction";
-export { default as ChurchHistory } from "./ui/ChurchHistory";
+
 export { default as ChurchStatistics } from "./ui/ChurchStatistics";
