+++ src/features/ministry/index.ts
@@ -1,0 +1,1 @@
+export { default as PastorIntroduction } from "./ui/PastorIntroduction";
